Yes
Python Dictionary:

Description: A dictionary in Python is an unordered collection of key-value pairs. It is defined using curly braces {}, with each key-value pair separated by a colon (:). Dictionaries are versatile data structures used to store and retrieve data quickly based on a unique key.
Python Modules:

Description: A module in Python is a file containing Python definitions and statements. It serves as a way to organize and reuse code. Modules can include variables, functions, and classes that can be imported into other Python scripts or modules. This promotes code modularity and reusability.